requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1
chromadb==0.4.22
sentence-transformers==2.3.1
streamlit==1.30.0
python-dotenv==1.0.0